fix exportMaterial
fix exportMaterial

+++ Project_Group4_SE1916/src/java/model/ExportDetail.java
@@ -1,25 +1,17 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
- */
 package model;
 
-/**
- *
- * @author ASUS
- */
 public class ExportDetail {
     private int exportDetailId;
     private int exportId;
     private int materialId;
-    private int quantity;
+    private double quantity; 
     private String materialCondition;
     private String reason;
 
     public ExportDetail() {
     }
 
-    public ExportDetail(int exportDetailId, int exportId, int materialId, int quantity, String materialCondition, String reason) {
+    public ExportDetail(int exportDetailId, int exportId, int materialId, double quantity, String materialCondition, String reason) {
         this.exportDetailId = exportDetailId;
         this.exportId = exportId;
         this.materialId = materialId;
@@ -27,8 +19,6 @@ public class ExportDetail {
         this.materialCondition = materialCondition;
         this.reason = reason;
     }
-    
-    
 
     public int getExportDetailId() {
         return exportDetailId;
@@ -54,11 +44,11 @@ public class ExportDetail {
         this.materialId = materialId;
     }
 
-    public int getQuantity() {
+    public double getQuantity() {
         return quantity;
     }
 
-    public void setQuantity(int quantity) {
+    public void setQuantity(double quantity) {
         this.quantity = quantity;
     }
 
@@ -77,7 +67,4 @@ public class ExportDetail {
     public void setReason(String reason) {
         this.reason = reason;
     }
-
-    
 }
-
